+++ data/core/statusview.lua
@@ -39,6 +39,17 @@ function StatusView:show_message(icon, icon_color, text)
 end
 
 
+function StatusView:show_tooltip(text)
+  self.message = { style.text, text }
+  self.message_timeout = system.get_time() + 1000
+end
+
+
+function StatusView:remove_tooltip()
+  self.message_timeout = 0
+end
+
+
 function StatusView:update()
   self.size.y = style.font:get_height() + style.padding.y * 2
 
@@ -94,6 +105,9 @@ function StatusView:get_items()
     local dv = core.active_view
     local line, col = dv.doc:get_selection()
     local dirty = dv.doc:is_dirty()
+    local indent = dv.doc.indent_info
+    local indent_label = indent.type == "hard" and "tabs: " or "spaces: "
+    local indent_size = tostring(indent.size) .. (indent.confirmed and "" or "*")
 
     return {
       dirty and style.accent or style.text, style.icon_font, "f",
@@ -108,6 +122,8 @@ function StatusView:get_items()
       self.separator,
       string.format("%d%%", line / #dv.doc.lines * 100),
     }, {
+      style.text, indent_label, indent_size,
+      style.dim, self.separator2, style.text,
       style.icon_font, "g",
       style.font, style.dim, self.separator2, style.text,
       #dv.doc.lines, " lines",