+++ data/plugins/detectindent.lua
@@ -1,8 +1,10 @@
 local core = require "core"
 local command = require "core.command"
+local common = require "core.common"
 local config = require "core.config"
 local DocView = require "core.docview"
 local Doc = require "core.doc"
+local tokenizer = require "core.tokenizer"
 
 local cache = setmetatable({}, { __mode = "k" })
 
@@ -42,12 +44,43 @@ local function optimal_indent_from_stat(stat)
   return bins[1][1], bins[1][2]
 end
 
-local auto_detect_max_lines = 400
+
+-- return nil if it is a comment or blank line or the initial part of the
+-- line otherwise.
+-- we don't need to have the whole line to detect indentation.
+local function get_first_line_part(tokens)
+  local i, n = 1, #tokens
+  while i + 1 <= n do
+    local ttype, ttext = tokens[i], tokens[i + 1]
+    if ttype ~= "comment" and ttext:gsub("%s+", "") ~= "" then
+      return ttext
+    end
+    i = i + 2
+  end
+end
+
+local function get_non_empty_lines(syntax, lines)
+  return coroutine.wrap(function()
+    local tokens, state
+    local i = 0
+    for _, line in ipairs(lines) do
+      tokens, state = tokenizer.tokenize(syntax, line, state)
+      local line_start = get_first_line_part(tokens)
+      if line_start then
+        i = i + 1
+        coroutine.yield(i, line_start)
+      end
+    end
+  end)
+end
+
+
+local auto_detect_max_lines = 200
 
 local function detect_indent_stat(doc)
   local stat = {}
   local tab_count = 0
-  for i, text in ipairs(doc.lines) do
+  for i, text in get_non_empty_lines(doc.syntax, doc.lines) do
     local str = text:match("^ %s+%S")
     if str then add_to_stat(stat, #str - 1) end
     local str = text:match("^\t+")
@@ -58,19 +91,27 @@ local function detect_indent_stat(doc)
   table.sort(stat, function(a, b) return a[1] < b[1] end)
   local indent, score = optimal_indent_from_stat(stat)
   if tab_count > score then
-    core.log_quiet("Detect-indent: using tabs indentation")
-    return "hard"
-  elseif indent then
-    core.log_quiet("Detect-indent: using indentation size of %d", indent)
-    return "soft", indent
+    return "hard", config.indent_size, tab_count
+  else
+    return "soft", indent or config.indent_size, score or 0
   end
 end
 
 
+local doc_text_input = Doc.text_input
+local adjust_threshold = 4
+
 local function update_cache(doc)
-  local type, size = detect_indent_stat(doc)
-  if type then
-    cache[doc] = { type = type, size = size }
+  local type, size, score = detect_indent_stat(doc)
+  cache[doc] = { type = type, size = size, confirmed = (score >= adjust_threshold) }
+  doc.indent_info = cache[doc]
+  if score < adjust_threshold and Doc.text_input == doc_text_input then
+    Doc.text_input = function(self, ...)
+      doc_text_input(self, ...)
+      update_cache(self)
+    end
+  elseif score >= adjust_threshold and Doc.text_input ~= doc_text_input then
+    Doc.text_input = doc_text_input
   end
 end