+++ data/plugins/treeview.lua
@@ -23,7 +23,7 @@ function TreeView:new()
   TreeView.super.new(self)
   self.scrollable = true
   self.visible = true
-  self.init_size = true
+  self.init_size = config.treeview_size
   self.cache = {}
   self.last = {}
 end
@@ -184,12 +184,15 @@ end
 
 function TreeView:update()
   -- update width
-  local dest = self.visible and config.treeview_size or 0
   if self.init_size then
-    self.size.x = dest
-    self.init_size = false
-  else
-    self:move_towards(self.size, "x", dest)
+    self.size.x = self.init_size
+    self.init_size = nil
+  elseif self.goto_size then
+    if self.goto_size ~= self.size.x then
+      self:move_towards(self.size, "x", self.goto_size)
+    else
+      self.goto_size = nil
+    end
   end
 
   TreeView.super.update(self)
@@ -251,12 +254,37 @@ end
 -- init
 local view = TreeView()
 local node = core.root_view:get_active_node()
-node:split("left", view, true)
+local treeview_node = node:split("left", view, {x = true}, true)
+
+-- The toolbarview plugin is special because it is plugged inside
+-- a treeview pane which is itelf provided in a plugin.
+-- We therefore break the usual plugin's logic that would require each
+-- plugin to be independent of each other. In addition it is not the
+-- plugin module that plug itself in the active node but it is plugged here
+-- in the treeview node.
+local toolbar_plugin, ToolbarView = core.try(require, "plugins.toolbarview")
+if config.toolbarview ~= false and toolbar_plugin then
+  local toolbar_view = ToolbarView()
+  treeview_node:split("down", toolbar_view, {y = true})
+  command.add(nil, {
+    ["toolbar:toggle"] = function()
+      toolbar_view:toggle_visible()
+    end,
+  })
+end
+
 
 -- register commands and keymap
 command.add(nil, {
   ["treeview:toggle"] = function()
-    view.visible = not view.visible
+    if view.visible then
+      view.previous_size = view.size.x
+      view.visible = false
+      view.goto_size = 0
+    else
+      view.visible = true
+      view.goto_size = view.previous_size
+    end
   end,
 })